v0.04.9
Putting this up for now, as I need to go for awhile. The main class in timeHandler is being called in HMJGUI now, but there's some kind of exception happening; this likely has something to do with the foor loop. Note that the IF statement in the FOR LOOP (if (hour.equals("17"))) is currently commented out; this will not be the case once we have finished our app.